Clean up

+++ include/core/server.hpp
@@ -8,9 +8,9 @@
 
 class Server {
 public:
-    Server(const std::string& host, uint16_t port);
+    Server(const std::string& host, int port);
     std::string getAddress() const;
-    uint16_t getPort() const;
+    int getPort() const;
     std::string getId() const;
     bool isHealthy() const;
     void setHealthStatus(bool status);
@@ -23,24 +23,24 @@ public:
     std::chrono::system_clock::time_point getLastHealthCheckTime() const;
     void setLastHealthCheckTime(std::chrono::system_clock::time_point t);
     
-    uint64_t getRequestCount() const { return request_count_.load(); }
+    int getRequestCount() const { return request_count_.load(); }
     void incrementRequestCount() { request_count_.fetch_add(1); }
     
     void incrementActiveConnections();
     void decrementActiveConnections();
-    uint64_t getActiveConnections() const;
+    int getActiveConnections() const;
 
     void setProcess(std::unique_ptr<Process> proc) { process_ = std::move(proc); }
     Process* getProcess() const { return process_.get(); }
 
 private:
     std::string host_;
-    uint16_t port_;
+    int port_;
     bool is_healthy_;
     std::string id_;
     std::chrono::system_clock::time_point last_health_check_time_;
-    std::atomic<uint64_t> request_count_{0};
-    std::atomic<uint64_t> active_connections_{0};
+    std::atomic<int> request_count_{0};
+    std::atomic<int> active_connections_{0};
     std::unique_ptr<Process> process_;
     double cpu_usage;
     double memory_usage;
