+++ src/strategies/least_connections.cpp
@@ -2,15 +2,14 @@
 
 std::shared_ptr<Server> LeastConnectionsStrategy::selectServer(
     const std::vector<std::shared_ptr<Server>>& servers,
-    const loadbalancer::Request& request) 
-{
+    const loadbalancer::Request& request) {
     std::shared_ptr<Server> best_server = nullptr;
-    uint64_t min_connections = UINT64_MAX;
+    int min_connections = INT_MAX;
 
     for (const auto& server : servers) {
         if (!server->isHealthy()) continue;
 
-        uint64_t active_connections = server->getActiveConnections();
+        int active_connections = server->getActiveConnections();
         if (active_connections < min_connections) {
             min_connections = active_connections;
             best_server = server;
